requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
discord==1.7.3
discord.py==1.7.3
Flask==2.0.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
multidict==5.1.0
pycparser==2.20
PyNaCl==1.4.0
python-dotenv==0.18.0
six==1.16.0
typing-extensions==3.10.0.0
Werkzeug==2.0.1
yarl==1.6.3
youtube-dl==2021.6.6
replit==3.2.4
